javadoc comments, internal, namechange
